feat(condo): DOMA-8507 rollback address-service changes
